refactor(routes): moved guard routes from routes/ to routes/guard for better organization
