+++ client/src/routes/guards/AnonymousRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../context/authContext';
+import { useAuth } from '../../context/authContext';
 
 const AnonymousRoute = ({ children }) => {
   const { isAuthenticated, user, isLoading } = useAuth();
