+++ client/src/routes/guards/protectedRoutes.jsx
@@ -1,7 +1,7 @@
 // src/routes/ProtectedRoute.jsx
 import React from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { useAuth } from '../context/authContext';
+import { useAuth } from '../../context/authContext';
 
 const ProtectedRoute = ({ allowedRoles = [], children }) => {
   const { isAuthenticated, user, isLoading } = useAuth();
